simplify classes: Plant & PlantGrid

+++ src/simulation/world/World.java
@@ -8,7 +8,7 @@ public class World {
 	//CONSTS
 	public static final Color NOTHING_COLOR = Color.WHITE;
 	public static final Color WALL_COLOR = Color.GRAY;
-	public static final int SIZE = 500;
+	public static final double SIZE = 500;
 	public static final int NUMBER_OF_STARTING_CREATURES = 50;
 	//ATTRIBUTES
 	private final ArrayList<Creature> creatures;
